+++ examples/ex1.jl
@@ -4,6 +4,8 @@ using CPLEX
 # using Cbc
 
 include("solver.jl")
+include("2qubit_gates.jl")
+include("3qubit_gates.jl")
 
 #-------------------------------#
 #      User-defined inputs      #
@@ -36,7 +38,7 @@ params = Dict{String, Any}(
 "depth" => 5,
 
 # "elementary_g/ates" => ["U3", "cnot_12", "Identity"],
-# "elementary_gates" => ["T1", "T2", "T3", "H3", "cnot_12", "cnot_13", "cnot_23", "Tdagger2", "Tdagger3"],
+# "elementary_gates" => ["T1", "T2", "T3", "H3", "cnot_12", "cnot_13", "cnot_23", "Tdagger2", "Tdagger3", "Identity"],
 "elementary_gates" => ["S1", "S2", "H1", "H2", "cnot_12", "cnot_21", "Identity"],
 
 "target_gate" => QCO.MGate(),