+++ examples/run_examples.jl
@@ -21,7 +21,7 @@ test_gates = ["test_hadamard",
                "test_W_using_HCnot",
                "test_RX_on_q3"]
 
-# test_gates = ["test_cnot_21_with_U"]
+# test_gates = ["test_toffoli"]
 
 #----------------------------------------------#
 #      Quantum Circuit Optimization model      #
@@ -31,6 +31,8 @@ result_qcm = Dict{String,Any}()
 for gates in test_gates 
     params = getfield(Main, Symbol(gates))()
     qcm_optimizer = get_solver(params)  
-    global result_qcm = QCO.run_QCModel(params, qcm_optimizer, model_type="compact_formulation")
+    global result_qcm = QCO.run_QCModel(params, 
+                                        qcm_optimizer, 
+                                        model_type = "compact_formulation")
 end